+++ asf_search/__init__.py
@@ -14,8 +14,11 @@ except PackageNotFoundError as e:
 from .ASFSession import ASFSession
 from .ASFProduct import ASFProduct
 from .ASFSearchResults import ASFSearchResults
+from .ASFSearchOptions import ASFSearchOptions, validators
 from .exceptions import *
 from .constants import *
 from .health import *
 from .search import *
 from .download import *
+from .CMR import *
+from .baseline import *