+++ CHANGELOG.md
@@ -1,5 +1,116 @@
 # Changelog
 
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/) 
+and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+
+<!--
+## Example template!!
+
+## [version](https://github.com/asfadmin/Discovery-PytestAutomation/compare/vOLD...vNEW)
+
+### Added:
+-
+
+### Changed:
+-
+
+### Fixed:
+- 
+
+### Removed:
+-
+
+-->
+
+## [4.0.0](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.4...v4.0.0)
+### Added
+- `ASFSearchOptions`: This class provides a number of useful ways to build search results
+  - Search parameters are immediately validated upon object creation/edit instead of at search time, which should lead to fewer errors at search time
+  - All search functions allow both the previous style of keyword arguments, as well as simply passing in an ASFSearchOptions object using the `opts` keyword arg. `opts` is always optional.
+    - If both approaches are used, the two are merged, with specific keyword args superseding the options in the object
+    - Most search functions now expect only their specific parameters, and an optional `opts` parameter. This allows simple usage in most cases, while the `opts` parameter provides access to advanced behavior or alternate workflows.
+  - Internally, all search functions work by passing ASFSearchOptions objects. This allows consistency when working with differently-configured search environments, such as in development.
+  - `ASFSearchResults` objects now include a `searchOptions` property, which describes the search used to create those results. This object can be copied, altered, used for subsequent searches, etc.
+    - When downloading, `ASFSearchResults` and `ASFProduct` default to use the session inside `searchOptions`, so you don't have to pass the same session in for both fetching and downloading results.
+- Exposed `get_stack_opts()` to support more approaches for building insar stacks.
+  - `get_stack_opts()` accepts an `ASFProduct` as a stack reference and returns the ASFSearchOptions object that would be used to build a corresponding insar stack
+    - A matching convenience method has been added to `ASFProduct`
+  - Supports the new `opts` argument described above.
+
+### Changed
+- All search functions now accepts the optional `opts=` argument, see `ASFSearchOptions` notes above.
+- Replaced all `cmr_token` key arguments with `session`, which takes a `Session`-compatible object. See https://docs.asf.alaska.edu/asf_search/ASFSession/ for more details.
+- Removed old GitHub actions
+
+------
+## [3.2.2](https://github.com/asfadmin/Discovery-PytestAutomation/compare/v3.2.1...v3.2.2)
+### Fixed
+- netrc authentication works again, affects `ASFProduct.download()`, `ASFSearchResults.download()`, `download_urls()`, `download_url()`
+
+------
+## [3.2.1](https://github.com/asfadmin/Discovery-PytestAutomation/compare/v3.2.0...v3.2.1)
+### Fixed
+- `ASFProduct.stack()` and `asf_search.baseline_search.stack_from_id()` now return ASFSearchResults instead of a list
+
+------
+## [3.2.0](https://github.com/asfadmin/Discovery-PytestAutomation/compare/v3.1.3...v3.2.0)
+### Changed
+- `ASFProduct.stack()` and `asf_search.baseline_search.stack_from_id()` now calculate `temporalBaseline` and `perpendicularBaseline` values of stacked products locally
+- `search()` now internally uses a custom format when communicating with ASF's SearchAPI. This should have no apparent impact on current usage of asf_search. 
+
+------
+## [3.1.3](https://github.com/asfadmin/Discovery-PytestAutomation/compare/v3.1.2...v3.1.3)
+### Fixed
+- Centroid calculation fixed for scenes spanning the antimeridian
+
+------
+## [3.1.2](https://github.com/asfadmin/Discovery-PytestAutomation/compare/v3.1.1...v3.1.2)
+### Changed
+- `ASFSession` methods `auth_with_cookiejar()` and `auth_with_token()` now raise an error if the passed cookiejar/token is invalid or expired
+- `ASFAuthenticationError` raised when encountering a 400 level error while downloading files
+### Fixed
+- Downloading files with sessions authenticated by `auth_with_token()` method works again
+
+------
+## [3.1.1](https://github.com/asfadmin/Discovery-PytestAutomation/compare/v3.1.0...v3.1.1)
+### Fixed:
+- Fixes missing CMR module import
+
+------
+## [3.1.0](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.6...v3.1.0)
+### Added
+- Added walkthrough in the form of several jupyter notebooks in /examples
+- Added `campaigns()` in `Campaigns` module, returns a list of campaigns for `UAV, AIRSAR, SENTINEL-1 INTERFEROGRAM (BETA)` platforms
+
+### Changed
+- Re-enable run-pytest workflow
+  - Add tests for `ASFSearch, ASFSession, ASFProduct` as well as baseline, geographic, and search modules
+  - Add Pytest-Automation Plugin integration
+  - Add automated CodeCov badge to readme
+- "collectionName" parameter in `geo_search()` and `search()` is deprecated and raises a warning. Will be removed in a future release, use "campaign" instead
+
+### Fixed
+- Fix error while raising ASFBaselineError in `baseline_search.get_stack_params()`
+
+------
+## [3.0.6](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.5...v3.0.6)
+### Changed
+- Skip download if file already exists
+  - In the future we will apply file size and/or checksum checks to ensure the existing file is correct
+
+------
+## [3.0.5](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.4...v3.0.5)
+### Added
+- Add documentation URL to setup.py
+- Add Gitter badge/link to readme
+### Fixed
+- Change hyphens to underscores in some product type constants
+
+------
 ## [3.0.4](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.3...v3.0.4)
 ### Changed
 - When working with source, package **must** be installed directly:
@@ -7,6 +118,7 @@
 ### Fixed
 - In-region S3 downloads should now function without issue
 
+------
 ## [3.0.3](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.2...v3.0.3)
 ### Fixed
 - Replace `ASFProduct.centroid()` calculation with shapely-based calculation
@@ -14,6 +126,7 @@
   - Removes numpy requirement
   - Adds shapely requirement
 
+------
 ## [3.0.2](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.0...v3.0.2)
 ### Added
 - Feature and Bug Report github issue templates
@@ -22,6 +135,7 @@
 - Fix download authentication header issue during direct-to-S3 redirects
 - Fix Sentinel-1 stacking to include both A and B in stacks
 
+------
 ## [3.0.0](https://github.com/asfadmin/Discovery-asf_search/compare/v2.0.2...v3.0.0)
 ### Added
 - Auth support for username/password and cookiejars, in addition to the previously available token-based approach. Create a session, authenticate it with the method of choice, then pass the session to whichever download method is being used.
@@ -40,19 +154,23 @@
 - `download_url()`, `download_urls()`, `ASFProduct.download()` and `ASFSearchResults.download()` now expect a `session` argument instead of `token`
 - Send auth headers to every step along a download redirect chain (including final AWS S3 buckets)
 
+------
 ## [2.0.2](https://github.com/asfadmin/Discovery-asf_search/compare/v2.0.1...v2.0.2)
 ### Added
 - INSTRUMENT constants for C-SAR, PALSAR, and ANVIR-2
 
+------
 ## [2.0.1](https://github.com/asfadmin/Discovery-asf_search/compare/v2.0.0...v2.0.1)
 ### Fixed
 - Versioning workflow corrected for proper versioning, stop bumping major instead of patch!
 
+------
 ## [2.0.0](https://github.com/asfadmin/Discovery-asf_search/compare/v1.1.0...v2.0.0)
 ### Fixed
 - Fixed import order of operations bug
 - Updated ASFProduct and ASFSearchResults to use path arg in download methods
 
+------
 ## [1.1.0](https://github.com/asfadmin/Discovery-asf_search/compare/v0.4.0...v1.1.0)
 ### Added
 - Parallel downloads now supported by ASFSearchResults. Defaults to 1 (sequential download)
@@ -66,6 +184,7 @@
 - Fixed ASFProduct import in search.py
 - importlib metadata fix for python <3.8
 
+------
 ## [0.4.0](https://github.com/asfadmin/Discovery-asf_search/compare/v0.3.0...v0.4.0)
 ### Added
 - ASFSearchResults now has a geojson() method which returns a data structure that matches the geojson specification
@@ -88,7 +207,7 @@
 ### Fixed
 - Cleaned up cruft from various refactors
 
-
+------
 ## [0.3.0](https://github.com/asfadmin/Discovery-asf_search/compare/v0.2.4...v0.3.0)
 
 ### Added
@@ -116,6 +235,7 @@
 - unused import cleanup
 - better type hinting on centroid() function
 
+------
 ## [0.2.4](https://github.com/asfadmin/Discovery-asf_search/compare/v0.0.0...v0.2.4)
 
 ### Added
@@ -135,3 +255,5 @@
 ### Fixed
 - Removed hard-coded version string
 - Install setuptools_scm in pypi publish action
+
+------